Check if .rocks exists

+++ test/conftest.py
@@ -218,7 +218,8 @@ def project_getting_started(cartridge_cmd, short_tmpdir):
     def create_getting_started_app(basepath):
         path = os.path.join(basepath, name)
         shutil.copytree(getting_started_path, path)
-        shutil.rmtree(os.path.join(path, '.rocks'))
+        if os.path.exists(os.path.join(path, '.rocks')):
+            shutil.rmtree(os.path.join(path, '.rocks'))
         shutil.rmtree(os.path.join(path, 'tmp'))
         return path
